+++ checkpoint/orbax/checkpoint/__init__.py
@@ -20,6 +20,7 @@ import functools
 
 import nest_asyncio
 from orbax.checkpoint import aggregate_handlers
+from orbax.checkpoint import args
 from orbax.checkpoint import checkpoint_utils
 from orbax.checkpoint import msgpack_utils
 from orbax.checkpoint import test_utils